+++ packages/gov-ui-kit/.storybook/manager.ts
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/manager-api';
+import aragonOdsTheme from './theme';
+
+addons.setConfig({
+    theme: aragonOdsTheme,
+});