+++ packages/gov-ui-kit/.storybook/theme.ts
@@ -1,0 +1,24 @@
+import { create } from '@storybook/theming';
+
+export default create({
+    base: 'light',
+    brandTitle: 'Aragon ODS',
+    brandUrl: 'https://aragon.org',
+    brandImage: 'https://i.postimg.cc/RVVSGThD/logo.png',
+    brandTarget: '_blank',
+    colorPrimary: '#003bf5',
+    colorSecondary: '#3164fa',
+    appBg: '#e4e7eb',
+    appContentBg: '#ffffff',
+    appBorderColor: '#cbd2d9',
+    appBorderRadius: 4,
+    textColor: '#323f4b',
+    textInverseColor: '#ffffff',
+    barTextColor: '#323f4b',
+    barSelectedColor: '#003bf5',
+    barBg: '#e4e7eb',
+    inputBg: '#ffffff',
+    inputBorder: '#cbd2d9',
+    inputTextColor: '#323f4b',
+    inputBorderRadius: 4,
+});